+++ runelite-client/src/main/java/net/runelite/client/plugins/grounditems/GroundItem.java
@@ -39,13 +39,13 @@ class GroundItem
 	private int quantity;
 	private WorldPoint location;
 	private int height;
-	private int haPrice;
+	private long haPrice;
 	private int gePrice;
 	private int offset;
 	private boolean tradeable;
 	private boolean isMine;
 
-	int getHaPrice()
+	long getHaPrice()
 	{
 		return haPrice * quantity;
 	}
